requirements.txt
# CategoryRAG系统完整依赖包

# 核心RAG依赖
chromadb>=0.4.0
sentence-transformers>=2.2.0
pyyaml>=6.0

# LLM API
openai>=1.0.0
dashscope>=1.14.0  # 阿里云Qwen API

# 文档处理工作流程依赖包
markitdown[docx]>=0.0.1a2

# Excel处理
pandas>=1.5.0
openpyxl>=3.0.0

# 表格格式化
tabulate>=0.9.0

# 向量和机器学习
numpy>=1.21.0
scikit-learn>=1.0.0

# HTTP请求
requests>=2.28.0

# 日志和工具
pathlib2>=2.3.0; python_version < "3.4"

# 可选：本地BGE模型支持
# FlagEmbedding>=1.2.0
# torch>=1.13.0
